moved selectors to common
